fifth commit

+++ internal/handlers/withdrawal/api/get_balance.go
@@ -1,6 +1,6 @@
 package api
 
 type GetBalanceResponse struct {
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
