+++ internal/handlers/withdrawal/api/get_withdrawals_info.go
@@ -8,6 +8,6 @@ type GetWithdrawalsInfoResponse struct {
 
 type Withdrawal struct {
 	Order       string    `json:"order"`
-	Sum         int       `json:"sum"`
+	Sum         float32   `json:"sum"`
 	ProcessedAt time.Time `json:"processed_at"`
 }
