+++ internal/handlers/withdrawal/api/make_withdrawal.go
@@ -1,6 +1,6 @@
 package api
 
 type MakeWithdrawalRequest struct {
-	Order string `json:"order"`
-	Sum   int    `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float32 `json:"sum"`
 }
