+++ internal/repository/migrations/010425_create_table_users_and_sessions.sql
@@ -5,7 +5,7 @@ create table if not exists users
     id       varchar(50) primary key,
     login    varchar(50) not null unique,
     password text        not null,
-    balance  int default 0
+    balance  numeric(12, 2) default 0
 );
 
 CREATE INDEX users_on_login_password_idx ON users (login, password);
