+++ internal/repository/migrations/150425_create_table_orders_and_withdrawls.sql
@@ -1,11 +1,12 @@
 -- +goose Up
 -- +goose StatementBegin
-CREATE TYPE order_status_type AS ENUM ('UNKNOWN', 'REGISTRED', 'INVALID', 'PROCESSING', 'PROCESSED');
+CREATE
+TYPE order_status_type AS ENUM ('UNKNOWN', 'REGISTRED', 'INVALID', 'PROCESSING', 'PROCESSED');
 
 create table if not exists orders
 (
     id          varchar(50) primary key,
-    accrual     int,
+    accrual     numeric(12, 2),
     status order_status_type default 'UNKNOWN',
     uploaded_at TIMESTAMP   not null,
     user_id     varchar(50) not null,
@@ -15,8 +16,8 @@ create table if not exists orders
 create table if not exists withdrawals
 (
     order_id     varchar(50) primary key,
-    sum          int         not null,
-    processed_at TIMESTAMP   not null,
-    user_id      varchar(50) not null
+    sum          numeric(12, 2) not null,
+    processed_at TIMESTAMP      not null,
+    user_id      varchar(50)    not null
 );
 -- +goose StatementEnd
